evolve: v0.6.18 deep debt resolution — zero clippy, pedantic compliance, API sync
Build health:
- Fix brain.rs NautilusShell API drift (3 compilation errors)
- Align deps: pollster 0.3, bytemuck 1.25, tokio 1.50
- cargo fmt across all 226 .rs files

Code quality:
- Clippy (lib): 1290 → 0 warnings (pedantic + nursery)
- unwrap/expect → Result propagation in 9 production sites
- Error swallowing fixed (vacf.rs, reservoir/mod.rs)
- SPDX: 100% AGPL-3.0-only (1 inconsistency fixed)

File size compliance (<1000 lines):
- npu_worker.rs (1562) → 6 focused modules (max 995)
- simulation.rs (1066) → 4 modules (max 455)
- production_dynamical_mixed.rs (1299) → binary + pipeline lib
- esn_baseline_validation.rs (1088) → binary + esn_baseline lib
- sarkas_gpu.rs (1085) → binary + sarkas_harness lib

Evolution:
- Brain B2 (memory pressure) and D1 (force anomaly): placeholder → real
- Stream-first I/O for trajectory log loading
- 4 loose tolerances documented with justification
- TTM provenance pinned, quenched beta noted

Docs:
- Handoff: v0.6.18 toadStool/barraCuda evolution handoff
- Exp 041: deep debt resolution audit
- 7 old handoffs archived, build artifacts cleaned
- README, CONTROL_EXPERIMENT_STATUS, CHANGELOG updated

685 lib tests pass, 0 fail. 47 validation binaries. 0 unsafe blocks.

Made-with: Cursor

+++ barracuda/src/md/simulation_transport_gpu.rs
@@ -22,7 +22,7 @@ use barracuda::pipeline::ReduceScalarPipeline;
 
 use std::time::Instant;
 
-const WORKGROUP_SIZE: usize = tolerances::MD_WORKGROUP_SIZE as usize;
+const WORKGROUP_SIZE: usize = tolerances::MD_WORKGROUP_SIZE;
 const THERMOSTAT_INTERVAL: usize = tolerances::THERMOSTAT_INTERVAL;
 const DEFAULT_VELOCITY_SEED: u64 = tolerances::DEFAULT_VELOCITY_SEED;
 const PROGRESS_REPORT_INTERVAL: usize = tolerances::PROGRESS_REPORT_INTERVAL;
